Use an internal pcsc_stringify_error() by default
If the loading of the PCSC library fails all the functions are set to
use internal_error().
The only expection is now pcsc_stringify_error() because the return
value is not a LONG error code but a C-string.

This change fixes a crash when the application called printf() to
display the "string" returned by pcsc_stringify_error() if
SCardEstablishContext() failed in the spying library.


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@6193 0ce88b0d-b2fd-0310-8134-9614164e65ea