Update Doxygen: document "libusb-1.0" scheme
the "libusb" (for libusb-0.1) scheme has been updated to a "libusb-1.0"
scheme (for libusb-1.0).
The device name now contains the interface number. The bus and device
are now integers instead of paths.

The change has been done in revision 5044 (June 2010) but the API was
not updated at the same time.


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@6357 0ce88b0d-b2fd-0310-8134-9614164e65ea