Add a reader reference count to be used as a Garbage Collector
A reader should be removed only when no one is using it, not just when
the reader is unplugged.

An application may be in the middle of using the reader when the reader
is unplugged. If the reader is removed right when unplugged then pcscd
may use a no more existing reader structure and crash.

Thanks to Martin Vogt for the bug report.
http://archives.neohapsis.com/archives/dev/muscle/2012-q2/0041.html


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@6388 0ce88b0d-b2fd-0310-8134-9614164e65ea