Make hLockId field volatile
The ReaderContext structure is accessed from different pcscd threads. To
prevent the compiler to optimise read access to the hLockId field we
make it volatile.

That may be the source of a problem when a reader is removed but pcscd
do not behave correctly.

Thanks to Torsten Hilbrich for the idea.
http://archives.neohapsis.com/archives/dev/muscle/2012-q3/0086.html


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@6458 0ce88b0d-b2fd-0310-8134-9614164e65ea