RFCreateReaderHandle(): Unref the reader is the handle is already used
The function RFCreateReaderHandle() checks if a handle is already used
by calling RFReaderInfoById() to find the associated reader.

So if RFReaderInfoById() returns with success then the handle is valid
and can't be used. A new random handle must be generated.

But since RFReaderInfoById() called REF_READER() internally we must call
UNREF_READER() to decrement the reference counter.


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@6459 0ce88b0d-b2fd-0310-8134-9614164e65ea