Do not use pthread_atfork() any more
pthread_atfork() was used to invalidate all the handles after a fork.
The problem is that the handler function was still called after the
libpcsclite library was unloaded on FreeBSD. This caused a crash of the
application.

The invalidation of the handles is not mandatory. After a fork the
application may still use the PC/SC handles and may get a
SCARD_F_COMM_ERROR error if the handle has been closed and in the pcscd logs:
winscard_svc.c:908:MSGCheckHandleAssociation() Client failed to authenticate

Closes Alioth bug [#314078] Do not use pthread_atfork() in a library
https://alioth.debian.org/tracker/index.php?func=detail&aid=314078&group_id=30105&atid=410085


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@6680 0ce88b0d-b2fd-0310-8134-9614164e65ea