Fix compiler warning
atrhandler.c: In function ‘ATRDecodeAtr’:
atrhandler.c:88:19: warning: variable ‘TCi’ set but not used [-Wunused-but-set-variable]
atrhandler.c:88:14: warning: variable ‘TBi’ set but not used [-Wunused-but-set-variable]

We don't use TBi and TCi but we must calculate them because of the p++
in the formulae.

See revision 5808 to know why these values are important.


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@6701 0ce88b0d-b2fd-0310-8134-9614164e65ea