Add ReiserFS support (and maybe some other strange file systems)
"the attached patch fixes incorrect use of the dirent::d_type field. As
documented in readdir(3) d_type may be set to DT_UNKNOWN, which
applications must handle properly.

The effect of the bug is that config files are skipped in case the
config directory is located on a file system that doesn't support d_type
(e.g. ReiserFS)."

Thanks to Ingo Weinhold for the patch


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@6944 0ce88b0d-b2fd-0310-8134-9614164e65ea