Increase the thread stack size to at least 4 MB
" Recently I tried to run it on a non-glibc system using musl instead.
Turns out musl has a default stacksize of 80KB instead of the
multi-megabyte we are used to in glibc. This leads to segfaults when
pcscd tries to allocate buffers on the stack. To fix this i applied the
following patch:
https://github.com/stef/aports/blob/master/main/pcsc-lite/musl-stacksize.patch

I would be very thankful if you would incorporate this "upstream" so
that other musl users can happily enjoy pcsc and friends. "

Thanks to Stefan Marsiske for the patch


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@6948 0ce88b0d-b2fd-0310-8134-9614164e65ea