Revamp libudev hotplug
" Hi,

this is a complete revamp of src/hotplug_libudev.c hotplug handling.
This patch fix a lot of race conditions and make the code a bit cleaner.

The following benefits will be achieved:

- udev events will be handled seperatly, a "add" event will only add a
device and a "remove" will now only remove the device. So the whole
bookkeeping is now done by the udev daemon, there is no more need to
track the state of the device. Also it fix a race where udev had yet not
finished the plug event and the device is not full accessible.

- fix the race gap between the coldplug scan and the activating of the
monitoring hotplug events. Now the monitoring will be first activated
and the the coldplug scan will be done. So there is a small change that
a device will be added during the coldplug scan. For this the
HPAddDevice() function needs to check for duplicate adds of a device.
But that is common design of libudev usage.

- No more mutex needed since there is no more concurrency, because the
coldplug scan will be no done inside the thread.
 
- Do not start the thread detached since there is no way to do a
synchronized stop of the thread. The HPStopHotPluggables() will now wait
until the thread is gone. So it is save to free memory of structures
since no one is access them.

- Remove the cooperative thread cancellation by AraKiriHotPlug. Despite
the non political correctness it will not work as expected, because the
select() system call will block infinitly. Using the pthread_cancel()
mechanism will do a better and safer job.

- Add signal safety to the select() system call. A EINTR should not lead
in a termination of the thread.

- Fix a typo: pluggble > plugable

- Make less error prone for unexpected libudev return values. Check all
return values of udev_...() functions against NULL.

- Fix resource leak in hotplug thread. The udev_device_unref() was never
called on event "remove".

- Cleanup HPStopHotPluggables() function, which release now all
allocated resources, including udev. It also will not crash when in
HPRegisterForHotplugEvents() was not able to acquire resources.

- Replace the prehistoric select() by poll(), which is faster and
consumes less memory.

Changlog:

V1 4-Aug-2014
	first release
V2 5-Aug-2014
	Fix coding style. Optimize HPRemoveDevice()
	Cleanup HPStopHotPluggables()
	Fix a resource leak not calling udev_device_unref() in case of "add" or "remove" event.
V3 5-Aug-2014
	Fix resource leak in HPStopHotPluggables(), release udev
	Fix error handling in HPRegisterForHotplugEvents() and HPStopHotPluggables()
	Revert unneeded RFRemoveReader() in HPRegisterForHotplugEvents()

- Stefani

Signed-off-by: Stefani Seibold <stefani at seibold.net> "

Thanks to Stefani Seibold for the patch
http://lists.alioth.debian.org/pipermail/pcsclite-muscle/Week-of-Mon-20140804/000093.html


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@6951 0ce88b0d-b2fd-0310-8134-9614164e65ea