Create the libpcsclite/pcscd socket with the SOCK_CLOEXEC flag
The communication socket between libpcsclite and pcscd is created with
the SOCK_CLOEXEC flag added so the socket is automatically closed when
the process (or its childs) calls exec().

The problem happens when an application does exec() without cleaning its
use of libpcsclite using SCardReleaseContext(). I could not reproduce a
real problem but it has been reported on the field with a closed source
application.

Thanks to Stefani Seibold for the bug report and patch.


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@6978 0ce88b0d-b2fd-0310-8134-9614164e65ea