Make the license more 3-clause BSD like
The line "Changes to this license can be made only by the copyright
author with explicit written consent." was present in the pcsc-lite
license from the origin of the project (at least since the initial
version of COPYING in the VCS in March 2002).

This line is not present in the 3-clause BSD license
http://en.wikipedia.org/wiki/BSD_licenses

The change has been approved by David Corcoran, initial author of
pcsc-lite.

The change was requested by Oracle about the headers files winscard.h
and pcsclite.h " To include in OpenJDK, Oracle needs the right to
relicense under other terms."


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@7004 0ce88b0d-b2fd-0310-8134-9614164e65ea