Allow to use pcscd in a remote session
" Current implementation of polkit support disallows auth by 
challenge/response, effectively turning "auth_admin" present in the 
default policy to "no".

Attached patch makes possible to use polkit challenge/response agent to 
be authorized.

It makes possible to use pcscd in a remote session (providing admin 
password to the agent).

Possible problem: If the authorization agent is present and active, 
polkit_authority_check_authorization_sync() could take a long time (the 
time of users' response). If the next request comes in the same time, it 
is postponed until the previous one is handled. (Actions done by root 
are not postponed.) "

Thanks to Stanislav Brabec for the patch
http://lists.alioth.debian.org/pipermail/pcsclite-muscle/Week-of-Mon-20141201/000247.html


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@7050 0ce88b0d-b2fd-0310-8134-9614164e65ea