SCardTransmit() may return SCARD_E_INSUFFICIENT_BUFFER
SCardTransmit() now correctly returns SCARD_E_INSUFFICIENT_BUFFER when
pcbRecvLength is not big enough to receive the card response.

The APDU is succesfully exchanged with the card but the card response is
lost since the client application buffer is not large enough to receive
it.

Before this patch an error was reported in the CCID driver and the
application got a SCARD_E_NOT_TRANSACTED error instead.

Now the expected size is available in pcbRecvLength for the application
to adapt its buffer size and, possibly, recall SCardTransmit().

Thanks to Mironenko for the bug report
https://github.com/LudovicRousseau/CCID/pull/5