Threading: lock the PC/SC context in a safe way
Functions working on hContext now acquire the context lock while the
global thread lock is held.

This will avoid the context to be released between the context retrieval
and context lock. The code was ugly with a new call to
SCardGetContext()to check that the context was still valid.

Thanks to Luc Mazardo for the bug report.
http://lists.alioth.debian.org/pipermail/pcsclite-muscle/Week-of-Mon-20150608/000395.html