Threading: lock the card context in a safe way
Functions working on hCard now acquire the context lock while the
global thread lock is held.

This will avoid the context to be released between the context retrieval
and context lock. The code was ugly with a new call to
SCardGetContextAndChannelFromHandle() to check that the context was
still valid.

This patch similar than the one in
9681246742b812ba1b8acbe8892d8aa755215059 for hContext.

Thanks to Luc Mazardo for the bug report.
http://lists.alioth.debian.org/pipermail/pcsclite-muscle/Week-of-Mon-20150608/000395.html