SCardGetStatusChange(): fix card movement bug
The card movement counter eventCounter is a uint32_t and this value is
also storded in the upper 16 bits of dwEventState.

dwEventState should be copied in dwCurrentState by the application
before the next call to SCardGetStatusChange().

The upper 16 bits of dwCurrentState are then compared to eventCounter to
detect a card movement.

The alrogithm works fine if eventCounter can be stored on a 16-bits
integer. The function SCardGetStatusChange() will return immediatly if
eventCounter is greater than 2^16 even if no card movement happened.
This is because eventCounter will always be different to the lowest
16-bits of eventCounter stored in dwCurrentState.

The patch resets eventCounter to 0 after 0xFFFF so its value can always
be stored in a 16-bits integer.

Thanks to Alan for the bug report
https://alioth.debian.org/tracker/index.php?func=detail&aid=314929&group_id=30105&atid=410085
[#314929] logic error in the implementation of the READER_STATE.eventCounter feature