Initialize readerStates[].eventCounter to 0
The field readerStates[].eventCounter was not initialized. It was not a
real problem since the exact value is not used, only a value increment
is used.

The only problematic case was if readerStates[].eventCounter value is
greater than 2^16. In that case the function SCardGetStatusChange() is
no more able to detect card movement and would return immediatly.

See also the previous patch.

Thanks to Alan for the bug report.