Add support of remove and/or customize PC/SC reader names
To enable these two features you need to configure pcsc-lite with
--enable-filter.

Ignore some readers
===================
If the environment variable PCSCLITE_FILTER_IGNORE_READER_NAMES is
defined then it contains a list of patterns separated by the character
":".
If a pattern is found in a PC/SC reader name then this reader is ignored
and will not be reported by SCardListReaders() or any other PC/SC calls.

Extend reader names
===================
To differentiate the PC/SC reader names one idea is to use the host name
of the system. If the IT department is doing correctly his job every
laptop should have a different host name.

If the environment variable PCSCLITE_FILTER_EXTEND_READER_NAMES is
defined then it contains a string that will be added at the end of the
PC/SC reader names.
The computer host name is available in the variable $HOSTNAME. If you
want to have a space character between the PC/SC reader name and host
name you define PCSCLITE_FILTER_EXTEND_READER_NAMES as:
" $HOSTNAME".

See also
http://ludovicrousseau.blogspot.fr/2015/12/remove-andor-customize-pcsc-reader-names.html