Enable Trace and Profile features using compiler flags
" I have a couple of small suggestions regarding the tracing feature of
the PC/SC-Lite's client side library. This feature is controlled by
the "DO_TRACE" preprocessor definition.

First, it's impossible to trigger this feature using the compiler
flags: for some reason, the winscard_clnt.c file contains directive
"#undef DO_TRACE". If it were a commented "#define" directive, then
this would be more useful: the feature could be triggered both by
editing the source file and by modifying the compiler flags. "

The code now uses
//#define DO_TRACE
instead of

It is now possible to enable this feature using the compiler flag
-DDO_TRACE without modifying the source code.

Same thing for DO_PROFILE.

Thanks to Maksim Ivanov for the bug idea
http://lists.alioth.debian.org/pipermail/pcsclite-muscle/Week-of-Mon-20160307/000537.html