RFUnlockSharing: handle SCARD_SHARE_EXCLUSIVE
When a card is connected using SCARD_SHARE_EXCLUSIVE then
RFUnlockSharing() called by SCardEndTransaction() should not be able to
remove the exclusive access.

The bug was detected because the following sequence worked but should
not:
SCardConnect(..., SCARD_SHARE_EXCLUSIVE, ...);
SCardEndTransaction();

An error was reported only on the second SCardEndTransaction() call.

Now the first call to SCardEndTransaction() will fail and the card
connection will stay exclusive.

Thanks to Christophe Ferrando for the bug report in "[Pcsclite-muscle]
SCARD_E_NOT_TRANSACTED"
http://lists.alioth.debian.org/pipermail/pcsclite-muscle/Week-of-Mon-20160516/000598.html