SCardConnect(): Fix a possible duplicated hCard context
It was possible for 2 applications to call SCardConnect() at the same
time and that the same random hCard handle is generated and used for
the 2 applications.

The solution is to serialize the 2 generations of hCard handles. The
second randomly generated handle will be found as already used and
another handle will be generated.

Thanks to Maksim Ivanov for the bug report
"[Pcsclite-muscle] Possible generation of duplicate SCARDHANDLE"
http://lists.alioth.debian.org/pipermail/pcsclite-muscle/Week-of-Mon-20160801/000673.html