SYS_InitRandom(): new function to init the PRNG
The Pseudo Random Number Generator now has an independent function to
initialize its internal state.

The PRNG init was done in the first call to SYS_RandomInt().
A problem could appear when SYS_RandomInt() is called for the first time
at the same time from 2 clients.

Thanks to Maksim Ivanov for the idea
"[Pcsclite-muscle] Possible generation of duplicate SCARDHANDLE"
http://lists.alioth.debian.org/pipermail/pcsclite-muscle/Week-of-Mon-20160801/000686.html