SCardDisconnect(): much faster SCARD_UNPOWER_CARD
When SCardDisconnect() is used with SCARD_UNPOWER_CARD we just power
down the card and do not power it up again.

The idea is to be faster and gain some milliseconds.
The code to power up the card on demand was already present since 2010
with the card auto power on and off mechanism.

Before the change: SCardDisconnect(SCARD_UNPOWER_CARD) in 61 ms
After the change: SCardDisconnect(SCARD_UNPOWER_CARD) in 1.4 ms

Improvement factor (for the card used) = 61/1.4 = x44

If you use a card that is slower to power up then the gain is even higher.

Thanks to Christophe Ferrando from Sylyca for the patch.