SCardGetStatusChange(): Fix the fix for a race condition
The change in 4e2a563c8ed4353ad013de85b71aac12ec599f82 is not correct.

WRITE_BODY() is a macro composed of 2 lines of C so we can't use:
"if (test)
    WRITE_BODY()"

As is is interpreted as:
"if (test)
    instruction1;
instruction2;"

For that to work we need to:
- use { } around WRITE_BODY()
or
- define WRITE_BODY() as a C instruction

Defining WRITE_BODY() as a C instruction (a "do {} while(0)") is the
best option as it will prevent the same bug to appear again.

Thanks to Florian Kaiser (again) for the patch
"[Pcsclite-muscle] [PATCH] fix racecondition between winscard server and
clients"
http://lists.alioth.debian.org/pipermail/pcsclite-muscle/Week-of-Mon-20161205/000775.html