ContextThread(): fix a SCARD_CANCEL race bug
From Maksim Ivanov:
" It seems that there are still some issues with regard to the blocking
requests cancellation functionality in PC/SC-Lite:

1. Extra SCARD_E_CANCELLED events may be sent by the daemon to the
   client that was previously performing an SCardGetStatusChange call.

   One scenario is that two concurrent SCardCancel calls succeeding
   simultaneously, and another scenario is an SCardCancel call
   succeeding simultaneously with an event sending from the status
   handler thread.

   As a result, the app<->daemon communication will break.

Suggested solutions:

For #1, suggesting to change the SCARD_CANCEL handler to firstly do
   EHUnregisterClientForEvent, and, only if it returns success, then
   send the SCARD_E_CANCELLED event to the client. "

Now MSGSignalClient() is called only if EHUnregisterClientForEvent()
succeeds i.e. the client was still blocked in a SCardGetStatusChange().

If SCardGetStatusChange() was already canceled or finished then we do not
signal the client.

Thanks to Maksim Ivanov for the bug report and solution
"[Pcsclite-muscle] Data races related to SCardCancel"
http://lists.alioth.debian.org/pipermail/pcsclite-muscle/Week-of-Mon-20170116/000789.html