SCardControl() may return SCARD_E_INSUFFICIENT_BUFFER
SCardControl() now correctly returns SCARD_E_INSUFFICIENT_BUFFER when
pbRecvBuffer is not big enough to receive the card response.

The CCID driver does the check and correctly and returned
CM_IOCTL_GET_FEATURE_REQUEST in such a case but not all IFD Handler may
be smart enough.

This change is similar to the same change for SCardTransmit() done in
8eb9ea1b354b050f997d003cf3b0c5b56f29f9f7 (29 mai 2015)

Thanks to Maximilian Stein for the patch
"[Pcsclite-muscle] SCardControl() should return SCARD_E_INSUFFICIENT_BUFFER similar to SCardTransmit()"
http://lists.alioth.debian.org/pipermail/pcsclite-muscle/Week-of-Mon-20170213/000815.html