Fix SCardGetStatusChange() when a reader is removed
Because of a race condition, a smart card reader removal may, in some
conditions, not be detected on the client side.

The problem is that when a reader is removed the clients are notified
(using EHSignalEventToClients() function) two times:
- the first time in EHStatusHandlerThread() from eventhandler.c just
  before the event thread exists,
- the second time in removeReader() from readerfactory.c after the
  reader has been removed

The problem is that the first time the clients are notified the reader
state has not yet changed. The reader is still visible by the clients.
If a client is slow to register to new events it may miss the second
notification and then wait "forever" in SCardGetStatusChange()
("forever" is defined as 60 seconds here as it is the event loop timeout).

The solution is to removed the first (in fact useless) notification.

Thanks to Maximilian Stein for the bug report and patch
"[Pcsclite-muscle] Race condition with SCardGetStatusChange() when USB
Reader is removed"
http://lists.alioth.debian.org/pipermail/pcsclite-muscle/Week-of-Mon-20170508/000855.html