SCardDisconnect(): fix ATR update for SCARD_UNPOWER_CARD
When the card is disconnected with SCARD_RESET_CARD or
SCARD_UNPOWER_CARD then disable the current ATR early in the code and in
all cases.

The problem was that the ATR was disabled only in the case
SCARD_RESET_CARD and only in case of failure.
So if SCARD_UNPOWER_CARD was used the ATR was still considered as valid,
even if the card was no more powered.