SCardDisconnect(): fix status update for SCARD_UNPOWER_CARD
Store the value returned by IFDPowerICC(, IFD_POWER_DOWN, ).
This value is then used later to decide if the powering down was a
success or not and set the card status accordingly.

The problem was detected by pcsc_scan stress code. If a SCardTransmit()
fails because the card is removed then SCardDisconnect() is called to
unpower the card.
Because the card is removed the powering down fails but the card status
was not correctly updated.