RFAddReader: more correctly detect duplicate readers
To detect a duplicate reader, checking the reader name and port is,
sometimes, not enough.

The problem was when two readers are removed and re-inserted immediately.
One of the 2 readers may not yet be completely removed (from pcscd point
of view) when it is asked to insert it again.

Fixes #20
"Reader not detected when disconnecting / connecting usb hub"
https://github.com/LudovicRousseau/PCSC/issues/20