Do not (possibly) lock a reader if allocating hCard fails
In case of SCardConnect() the reader may be locked in
SCARD_SHARE_EXCLUSIVE mode if internal SCardConnect() works but
MSGAddHandle() fails because the list of handle is full.

You need to start pcscd with "--max-card-handle-per-reader n" with
n > 200 or the 200 limit (default value) will be hit in internal
SCardConnect() and MSGAddHandle() would not be called.

Thanks to Maksim Ivanov for the bug report
"[Pcsclite-muscle] SCardConnect behavior with invalid contexts"
http://lists.infradead.org/pipermail/pcsclite-muscle/2020-July/001095.html

" Hello,

I believe that there's a potential problem with the SCardConnect
implementation that it doesn't check the received SCARDCONTEXT
*before* executing the command. This might result in an unexpected
state, where the SCardConnect() caller receives an error code
meanwhile the connection to the card is actually established (which,
for example, might be an exclusive connection that prevents anyone
else from connecting to the card).

In detail, the ContextThread() function in winscard_svc.c, when
receiving the SCARD_CONNECT command, calls first SCardConnect() from
winscard.c, and then MSGAddHandle(). The former ignores SCARDCONTEXT
and, if possible, establishes a connection to the card. The latter
does check the SCARDCONTEXT value, but this happens after the
connection is already established, and its error is just returned to
the caller (without closing the just-opened connection).

Would it make sense to add a check of SCARDCONTEXT before calling
SCardConnect(), and/or to call SCardDisconnect() if MSGAddHandle()
fails?

Regards,
Maksim "