Update reader state when a card is removed during an exchange
If the smart card is removed during an APDU exchange we may receive
SCARD_E_NO_SMARTCARD. In this case we shall update the reader state to
indicate the card has been removed.

We shall do that *before* the next loop in EHStatusHandlerThread() so
that a call to SCardStatus() immediatly after the SCardTransmit() will
return the correct state.

This new mechanism is also implemented in the CCID driver version
1.4.34.

Thanks to Marc Kewitz for the bug report
"[Pcsclite-muscle] Race condition during readerstate update"
http://lists.infradead.org/pipermail/pcsclite-muscle/2020-August/001107.html