Use correct type for IFD* functions
Fix the ifdwrapper.h header file so the functions types correspond to
what we have in ifdwrapper.c since the commit
eecf6856b9d5c81ef4b3caa53cfc38d3075ca339

The compilation was broken on macOS with:
ifdwrapper.c:67:14: error: conflicting types for 'IFDSetPTS'
RESPONSECODE IFDSetPTS(READER_CONTEXT * rContext, DWORD dwProtocol,
             ^
./ifdwrapper.h:52:7: note: previous declaration is here
        LONG IFDSetPTS(READER_CONTEXT *, DWORD, UCHAR, UCHAR, UCHAR, UCHAR);
             ^

Compilation was correct on GNU/Linux but incorrect on macOS because we
have:
    typedef long RESPONSECODE;
For both macOS and GNU/Linux.

and
    typedef int32_t LONG;
for maCOS and
    typedef long LONG;
For GNU/LInux

So on macOS LONG and RESPONSECODE types are not equivalent.