hotplug libusb: support CCIDCLASSDRIVER
Try to use a (CCID) class driver if a specific driver fails to use the
reader.

This may happen if both acsccid and ccid drivers are installed.
acsccid should be used first.

This feature was present in src/hotplug_libudev.c but not yet in
src/hotplug_libusb.c.

" ccid vs acsccid - miss probing #26 "
https://github.com/acshk/acsccid/issues/26