SetProtocol: Handle IFD_NOT_SUPPORTED from the driver
If the driver returns IFD_NOT_SUPPORTED from the function
IFDHSetProtocolParameters() it indicates the "command" is not supported.
It is different from IFD_PROTOCOL_NOT_SUPPORTED that indicates the
"prorocol" is not supported.

The problem was with a ACR38U CCID and a memory card. The
SetParameters() fails but the card can still be used using the default
protocol.

The behaviour is identical to what we had before the commit
5d58577093f56dbcfbf754940f885ed8fcdba206.

Thanks to Godfrey Chung for the bug report
"pcsc-lite 1.9.1: memory card connection failed #103"
https://github.com/LudovicRousseau/PCSC/issues/103