hotplug_libusb: fix a memory leak
We must release the configuration descriptor even if the USB device is
NOT a smart card reader.

This problem was specific to hotplug using libusb. By default GNU/Linux
systems should use libudev instead. So the problem is visible in FreeBSD
systems or other non Linux systems.

Thanks to Sergey V. Koupreyenko for the bug report.

The problem was also reported to pfSense at
https://redmine.pfsense.org/issues/12095