Add tag TAG_IFD_DEVICE_REMOVED
This tag is used to tell the driver that the reader has been detected
removed. So the driver do not try to talk to the device since that will
fails.

The idea is to avoid error messages in the logs like:
00000000 [140611055253248] ccid_usb.c:871:WriteUSB() write failed (1/12): -4 LIBUSB_ERROR_NO_DEVICE
00000692 [140611271165696] ccid_usb.c:871:WriteUSB() write failed (1/12): -4 LIBUSB_ERROR_NO_DEVICE
when the device is removed and pcscd tells the driver to close the
connection.