pcscdaemon: autoexit even if no client connects
pcscd will now auto exit even if no PC/SC client has connected.
The idea is to auto exit if pcscd is restarted after an upgrade or a new
driver installation. In that case pcscd is NOT strated by a new client
so the alarm is not set the client disconnect.

If you use something like "service pcscd restart" to restart pcscd then
it will exit after 60 seconds (if no client uses it).