Fix variable substitution in systemd units
Before this change, Autoconf was used to subsitute variables, including
@sysconfdir@, in systemd unit files. This is not how Autoconf should be
used and had an unintended negative consequence: the @sysconfdir@
variable was not fully expanded, so literal `${prefix}/etc` was ending up
in the pcscd.service file, where it was invalid.

Use sed to perform the subsitution from within the makefile where the
variables get expanded fully. This seems to be the standard approach
recommended by Autoconf documentation and used in various projects.

Also, remove the `SCRIPT_IN_FILES` variable from the makefile, which was
not used and seems to be a copy-paste from some other project.